backend_django/pyproject.toml
[project]
name = "backend-django"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "channels>=4.3.2",
    "channels-redis>=4.3.0",
    "daphne>=4.2.1",
    "django>=5.2.12",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "djangorestframework-stubs>=3.16.8",
]
